python/requirements.txt
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-cloud-bigquery==2.23.2